requirements.txt
boto3==1.24.82
botocore==1.27.82
jmespath==1.0.1
numpy==1.23.3
pandas==1.5.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
